+++ backend/core/quivr_core/api/packages/files/parsers/epub.py
@@ -1,6 +1,6 @@
 from langchain_community.document_loaders.epub import UnstructuredEPubLoader
 
-from quivr_core.models.files import File
+from quivr_core.api.models.files import File
 
 from .common import process_file
