+++ backend/core/quivr_core/api/packages/files/parsers/github.py
@@ -5,7 +5,7 @@ from langchain.schema import Document
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_community.document_loaders import GitLoader
 
-from quivr_core.models.files import File
+from quivr_core.api.models.files import File
 from quivr_core.packages.embeddings.vectors import Neurons
 from quivr_core.packages.files.file import compute_sha1_from_content
