+++ backend/core/quivr_core/api/packages/files/parsers/html.py
@@ -1,6 +1,6 @@
 from langchain_community.document_loaders import UnstructuredHTMLLoader
 
-from quivr_core.models.files import File
+from quivr_core.api.models.files import File
 
 from .common import process_file
