+++ backend/core/quivr_core/api/packages/files/parsers/telegram.py
@@ -1,4 +1,4 @@
-from quivr_core.models.files import File
+from quivr_core.api.models.files import File
 from quivr_core.packages.files.loaders.telegram import TelegramChatFileLoader
 
 from .common import process_file
