+++ backend/core/quivr_core/api/packages/quivr_core/rag_factory.py
@@ -1,17 +1,17 @@
 from typing import Type
 
-from quivr_core.logger import get_logger
-from quivr_core.modules.brain.entity.integration_brain import IntegrationEntity
-from quivr_core.modules.brain.integrations.Big.Brain import BigBrain
-from quivr_core.modules.brain.integrations.GPT4.Brain import GPT4Brain
-from quivr_core.modules.brain.integrations.Multi_Contract.Brain import (
+from quivr_core.api.logger import get_logger
+from quivr_core.api.modules.brain.entity.integration_brain import IntegrationEntity
+from quivr_core.api.modules.brain.integrations.Big.Brain import BigBrain
+from quivr_core.api.modules.brain.integrations.GPT4.Brain import GPT4Brain
+from quivr_core.api.modules.brain.integrations.Multi_Contract.Brain import (
     MultiContractBrain,
 )
-from quivr_core.modules.brain.integrations.Notion.Brain import NotionBrain
-from quivr_core.modules.brain.integrations.Proxy.Brain import ProxyBrain
-from quivr_core.modules.brain.integrations.Self.Brain import SelfBrain
-from quivr_core.modules.brain.integrations.SQL.Brain import SQLBrain
-from quivr_core.modules.brain.knowledge_brain_qa import KnowledgeBrainQA
+from quivr_core.api.modules.brain.integrations.Notion.Brain import NotionBrain
+from quivr_core.api.modules.brain.integrations.Proxy.Brain import ProxyBrain
+from quivr_core.api.modules.brain.integrations.Self.Brain import SelfBrain
+from quivr_core.api.modules.brain.integrations.SQL.Brain import SQLBrain
+from quivr_core.api.modules.brain.knowledge_brain_qa import KnowledgeBrainQA
 
 logger = get_logger(__name__)
