+++ backend/core/quivr_core/api/packages/utils/telemetry.py
@@ -6,7 +6,7 @@ import threading
 import httpx
 from fastapi import Request
 
-from quivr_core.logger import get_logger
+from quivr_core.api.logger import get_logger
 
 logger = get_logger(__name__)
